Add an attribute set type based on typemap
This will replace our existing attribute system for one that’s a bit more generic. I’d quite like to be able to return all attributes implementing a trait or have a way to specify a ‘default’ trait that we can use when passing stuff into controllers but not sure how at the moment.